Устранение замечаний
